requirements.txt
python-telegram-bot>=20.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
yfinance>=0.2.36
alpha_vantage>=2.3.1
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1
urllib3<2.0.0  # Compatible with LibreSSL 